+++ NEWS.md
@@ -2,6 +2,8 @@
 
 # development version
 
+- Add Huh 2024 ritlecitinib ([doi:10.1007/s40262-024-01347-6](https://doi.org/10.1007/s40262-024-01347-6)) -- adolescents and adults with alopecia areata.
+
 - Add Ousey 2026 plozasiran ([doi:10.1002/jcph.70190](https://doi.org/10.1002/jcph.70190)) -- adults with familial chylomicronemia syndrome.
 
 - Add Ravva 2010 varenicline exposure-response models ([doi:10.1038/clpt.2009.282](https://doi.org/10.1038/clpt.2009.282)) -- adult cigarette smokers in smoking-cessation trials.